pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scouting-agent"
version = "0.1.0"
description = "Wyscout ingestion, Bruin/BigQuery pipeline, possession analysis"
requires-python = ">=3.11"
readme = "bruin-pipeline/README.md"
dependencies = [
  "pandas>=2.0",
  "python-dotenv>=1.0",
  "requests>=2.28",
  "tqdm>=4.65",
  "google-cloud-bigquery>=3.14",
  "google-cloud-storage>=2.14",
  "google-cloud-aiplatform>=1.70",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "numpy>=2.4.3",
  "matplotlib>=3.10.8",
  "mplsoccer>=1.6.1",
  "db-dtypes>=1.5.0",
  "langchain-core>=1.2.20,<2",
  "langchain-google-vertexai>=3.2.2,<4",
  "langchain-google-genai>=4.2.1",
  "langgraph>=1.1.3",
  "streamlit>=1.55.0",
]


[dependency-groups]
dev = [
  "ipython>=8.0",
  "ipykernel>=6.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.setuptools]
py-modules = [
  "wyscout",
  "possession_analyzer",
  "possession_qualifiers",
  "possession_description",
]
